Changes to Steering Actuation
Took the derivative between two torques and established a torque threshold based on the result. If the threshold was passed, it would change to a state that would keep the position of the wheels and steering the same while giving the user full control. Untested as of now, so purely theoretical at this time.